backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
fastf1
python-dotenv
google-generativeai